hackathon-2026/starter-kits/llm-basic/requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
# langchain (optional, uncomment to use)
# openai (optional, uncomment to use)
